+++ NEWS.md
@@ -1,6 +1,7 @@
 # mirai 0.7.0.9400 (development)
 
-* Allow user interrupts of `call_mirai()` again (regression in 0.7.0), returning a 'miraiError'.
+* Allow user interrupts of `call_mirai()` again (regression in 0.7.0), now returning a 'miraiInterrupt'.
+* Adds auxiliary function `is_mirai_interrupt()` to test if an object is a 'miraiInterrupt'.
 * Returned 'errorValues' e.g. mirai timeouts are no longer accompanied by a warning (change in nanonext 0.7.0).
 * Internal performance enhancements.